refactor(server): move pipeline references to their own module
